+++ src/components/DragBall.vue
@@ -1,7 +1,10 @@
 <!-- REF [Vue3+TS封装一个可任意拖拽的悬浮球组件，可拓展_codnan的博客-CSDN博客_vue3悬浮球](https://blog.csdn.net/annans/article/details/127427847) -->
 <script lang="ts" setup>
-import { ref } from "vue";
+import { ref, unref, watchEffect } from "vue";
+import { Storage } from "webextension-polyfill";
+
 import { Position } from "./../utils/position";
+import { browser } from "./../utils/browser";
 
 /* 初始位置 */
 const props = defineProps(["top", "bottom", "left", "right"]);
@@ -13,6 +16,36 @@ const left = ref<string | undefined>(props.left);
 const right = ref<string | undefined>(props.right);
 
 const position = new Position(); // 拖动前位置
+
+/* 浏览器扩展环境 */
+if (import.meta.env.PROD) {
+    /* 从储存中读取悬浮球位置 */
+    browser.storage.local
+        .get({
+            position: {
+                top: unref(top),
+                bottom: unref(bottom),
+                left: unref(left),
+                right: unref(right),
+            },
+        })
+        .then(items => {
+            if (top.value !== items.position.top) top.value = items.position.top;
+            if (bottom.value !== items.position.bottom) bottom.value = items.position.bottom;
+            if (left.value !== items.position.left) left.value = items.position.left;
+            if (right.value !== items.position.right) right.value = items.position.right;
+        });
+
+    /* 监听悬浮球位置更改 */
+    browser.storage.local.onChanged.addListener((changes: Storage.StorageAreaOnChangedChangesType) => {
+        if (changes.position) {
+            if (top.value !== changes.position.newValue.top) top.value = changes.position.newValue.top;
+            if (bottom.value !== changes.position.newValue.bottom) bottom.value = changes.position.newValue.bottom;
+            if (left.value !== changes.position.newValue.left) left.value = changes.position.newValue.left;
+            if (right.value !== changes.position.newValue.right) right.value = changes.position.newValue.right;
+        }
+    });
+}
 
 // 拖拽开始事件
 const dragstart = (e: DragEvent) => {
@@ -52,6 +85,18 @@ const dragend = (e: DragEvent) => {
         top.value = undefined;
         bottom.value = `${Math.max(ratio, 0)}vh`;
     }
+
+    /* 浏览器扩展环境 */
+    if (import.meta.env.PROD) {
+        browser.storage.local.set({
+            position: {
+                top: unref(top),
+                bottom: unref(bottom),
+                left: unref(left),
+                right: unref(right),
+            },
+        });
+    }
 };
 </script>
 
@@ -70,7 +115,7 @@ const dragend = (e: DragEvent) => {
 <style scoped lang="less">
 .drag-ball {
     cursor: pointer;
-    position: absolute;
+    position: fixed;
 
     // REF: [div中只有一张图片时，底部会有空白，需要加lineheight:0_扬州的你的博客-CSDN博客_div底部多出空白](https://blog.csdn.net/jtpython666/article/details/116278223)
     line-height: 0;