+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 - vue-i18n 打包时仅打包运行时模块 | vue-i18n bundle the runtime only module.
 - 调整界面样式 | Adjust interface style.
 - 用户配置数据持久化 | User configuration data persistence.
+- 关联搜索引擎 | Associate search engine.